requirements.txt
fastapi==0.63.0
sqlalchemy==1.3.23
beautifulsoup4==4.7.1
configparser==5.0.0
redis-py-cluster==2.0.0
cachetools==4.0.0
requests==2.22.0
pyjwt==1.7.1
pymongo[srv]==4.3.3
kafka-python==2.0.1
fastapi-utils==0.2.1
mongoengine==0.20.0
databases==0.3.2
loguru==0.5.2
aiosql==3.1.2
passlib==1.7.2
pydantic==1.7.3
pypika==0.39.1
slugify==0.0.1
uvicorn
jsonschema==3.2.0
pykafka==2.8.0
python-keycloak==0.24.0
itsdangerous==1.1.0
python-multipart==0.0.5
verify-email==2.4.3
nest-asyncio==1.5.8